backend/requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.6
h11==0.14.0
idna==3.10
pydantic==2.10.5
pydantic_core==2.27.2
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
python-dotenv==1.0.1
passlib==1.7.4
python-jose==3.3.0
alembic==1.13.1
httpx==0.27.0